backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
aiosqlite==0.20.0
python-multipart==0.0.9
openai==1.30.0
Pillow==10.3.0
colorthief==0.2.1
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.27.0
rembg[cpu]
onnxruntime
# ── Auth ──────────────────────────────────────────────────────────────────────
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
email-validator==2.1.1
aiosmtplib==3.0.1
# ── Security ───────────────────────────────────────────────────────────────────
slowapi==0.1.9
